requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==1.17.1
channels==4.2.2
constantly==23.10.4
cryptography==45.0.4
daphne==4.2.0
Django==5.2.2
djangorestframework==3.16.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
numpy==2.3.0
packaging==25.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
python-dotenv==1.1.0
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.14.0
zope.interface==7.2
